upload packs changes for console

+++ src/main/java/com/dpw/runner/shipment/services/utils/CommonUtils.java
@@ -3,6 +3,7 @@ package com.dpw.runner.shipment.services.utils;
 import com.dpw.runner.shipment.services.aspects.MultitenancyAspect.MultiTenancy;
 import com.dpw.runner.shipment.services.aspects.MultitenancyAspect.TenantContext;
 import com.dpw.runner.shipment.services.aspects.intraBranch.InterBranchContext;
+import com.dpw.runner.shipment.services.aspects.intraBranch.InterBranchContext;
 import com.dpw.runner.shipment.services.commons.constants.Constants;
 import com.dpw.runner.shipment.services.commons.constants.DaoConstants;
 import com.dpw.runner.shipment.services.commons.requests.Criteria;
@@ -539,8 +540,8 @@ public class CommonUtils {
 
     public void updateConsolOpenForAttachment(ConsolidationDetails consolidationDetails) {
         if(!Objects.isNull(consolidationDetails.getAchievedQuantities())) {
-            Integer weightUtilization = consolidationDetails.getAchievedQuantities().getWeightUtilization() != null ? Integer.valueOf(consolidationDetails.getAchievedQuantities().getWeightUtilization()) : 0;
-            Integer volumeUtilization = consolidationDetails.getAchievedQuantities().getVolumeUtilization() != null ? Integer.valueOf(consolidationDetails.getAchievedQuantities().getVolumeUtilization()) : 0;
+            Double weightUtilization = consolidationDetails.getAchievedQuantities().getWeightUtilization() != null ? Double.valueOf(consolidationDetails.getAchievedQuantities().getWeightUtilization()) : 0;
+            Double volumeUtilization = consolidationDetails.getAchievedQuantities().getVolumeUtilization() != null ? Double.valueOf(consolidationDetails.getAchievedQuantities().getVolumeUtilization()) : 0;
             if(weightUtilization > 100 || volumeUtilization > 100)
                 consolidationDetails.setOpenForAttachment(false);
         }
